Move DevicePool to its own file
